internal_dispose_GC should've used "int" for it's handle instead of "long"
git-svn-id: http://svn.warlock.cc/svn/warlock/warlock2/trunk@624 bb130017-240f-0410-9613-9c2c827820fb

+++ core/cc.warlock.rcp/src/cc/warlock/rcp/ui/WarlockText.java
@@ -572,11 +572,11 @@ public class WarlockText implements LineBackgroundListener, Drawable {
 		return textWidget.getContent();
 	}
 
-	public void internal_dispose_GC(long handle, GCData data) {
+	public void internal_dispose_GC(int handle, GCData data) {
 		textWidget.internal_dispose_GC(handle, data);
 	}
 
-	public long internal_new_GC(GCData data) {
+	public int internal_new_GC(GCData data) {
 		return textWidget.internal_new_GC(data);
 	}
 }
